.#1 merge movement fixes to the new Player script file
